moved new examples to demo-main, used same style as rest of the code
